SCLang: avoid collecting inside primitives (#7133)
Wait until the primitive has finished to call collect.
This has the potential draw back that if you are creating many temporary objects, you might run out of memory. But none of the primitives do this right now, and the ones that do allocate a lot of temporary objects usually disable the collection cycle anyway to avoid having the GC free the objects before they are placed on the stack.