SCLang: implement NaN tagging slotting (#6365)
This PR introduces one point of failure when reading doubles into slots from outside sclang, that is, if the double is a nan, and it happens to match one of the nan values use to do tagging, the slot will fail. In the best case, this will be undetectable but SCLang will behave oddly, in the worst case, this will segfault.

In future, make sure all doubles loaded into SCLang (from the network or files) pass the correct flag to the constructor, ensuring they are checked for safety.